monitor: Remove agent events queue
This commit avoids queueing agent events before sending them to the
monitor. The queue was introduced when the monitor was running in a
separate process to ensure that `SendEvent` would not block due to the
inter-process communication pipe not being ready.

In the current implementation however, sending an event to the monitor
already is a non-blocking operation: The listener `Enqueue` method
(which is invoked when sending an event to the monitor) already has a
queue per listener which ensures that `Enqueue` is non-blocking.

Therefore, there is no need to have an additional global queue.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>