ipcache: Keep port name conflicts
Do not warn of port name conflicts when ports are collected, but keep
the conficting ports. This allows (any) warnings to be postponed until
the named ports are used in a policy.

policy.NamedPortsMap is renamed to policy.NamedPortMap, and a new
policy.NamedPortMultiMap is defined, as well as an interface
policy.NamedPortsMap that abstracts the difference for callers of
GetNamedPort().

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>